IntLongMath,InconsistentCapitalization,UnnecessaryAssignment

+++ client-flink/flink-1.15/src/main/java/org/apache/celeborn/plugin/flink/RemoteShuffleResultPartition.java
@@ -202,7 +202,8 @@ public class RemoteShuffleResultPartition extends ResultPartition {
   public void updateStatistics(
       SortBuffer.BufferWithChannel bufferWithChannel, boolean isBroadcast) {
     numBuffersOut.inc(isBroadcast ? numSubpartitions : 1);
-    long readableBytes = bufferWithChannel.getBuffer().readableBytes() - BufferUtils.HEADER_LENGTH;
+    long readableBytes =
+        (long) bufferWithChannel.getBuffer().readableBytes() - BufferUtils.HEADER_LENGTH;
     numBytesProduced.inc(readableBytes);
     numBytesOut.inc(isBroadcast ? readableBytes * numSubpartitions : readableBytes);
   }
